Added dependencies for testing, added initial test
This initial commit for testing adds JUnit, Robolectric, Mockito, and PowerMock as dependencies for generating tests. The initial tests are on the Transaction class. Minification done by Proguard is currently disabled for debug builds to make it easier to add testing dependencies without modifying the Proguard file (although modifications have been made in the Proguard file to make it easier to re-enable in the future).